backend/requirements.txt
django
djangorestframework
gunicorn
uvicorn[standard]
psycopg2-binary
django-cors-headers
drf-spectacular
django-filter
